Refer PRG ROM from DMC directly
DMC triggers DMA to acquire sampling data from memory on hardware.
Currently emulate it using function callback.
But it will be prohibited on `AudioWorklet`
because they are running on different threads.

To make `DMC` run on `AudioWorklet`,
pass PRG ROM to DMC and refers it directly
according to bank change settings.